github actions demo

+++ src/components/Cart/RowCart.tsx
@@ -21,7 +21,9 @@ export default function RowCart({ eCart, index }: IRowCartProps) {
   // const summa = (eCart.count * (eCart?.price ?? 0)).toFixed(2);
 
   const onClickDeleteItemCart = () => {
-    dispatch(cartActions.removeItemCart({ id: eCart.good.id }));
+    if (window.confirm('Видалити?')) {
+      dispatch(cartActions.removeItemCart({ id: eCart.good.id }));
+    }
   };
 
   return (
